+++ weblate/trans/views.py
@@ -46,6 +46,7 @@ import datetime
 import logging
 import os.path
 import json
+import csv
 from xml.etree import ElementTree
 import urllib2
 
@@ -236,6 +237,63 @@ def upload_dictionary(request, project, lang):
     else:
         messages.error(request, _('Failed to process form!'))
     return HttpResponseRedirect(reverse('weblate.trans.views.show_dictionary', kwargs = {'project': prj.slug, 'lang': lang.code}))
+
+def download_dictionary(request, project, lang):
+    '''
+    Exports dictionary.
+    '''
+    prj = get_object_or_404(Project, slug = project)
+    lang = get_object_or_404(Language, code = lang)
+
+    # Parse parameters
+    export_format = None
+    if 'format' in request.GET:
+        export_format = request.GET['format']
+    if not export_format in ['csv', 'po']:
+        export_format = 'csv'
+
+    # Grab all words
+    words = Dictionary.objects.filter(project = prj, language = lang).order_by('source')
+
+    if export_format == 'csv':
+        response = HttpResponse(mimetype='text/csv; charset=utf-8')
+        response['Content-Disposition'] = 'attachment; filename=dictionary-%s-%s.csv' % (prj.slug, lang.code)
+
+        writer = csv.writer(response)
+
+        for word in words.iterator():
+            writer.writerow((word.source.encode('utf8'), word.target.encode('utf8')))
+
+        return response
+    elif export_format == 'po':
+        from translate.storage.po import pounit, pofile
+
+        response = HttpResponse(mimetype='text/x-po; charset=utf-8')
+        response['Content-Disposition'] = 'attachment; filename=dictionary-%s-%s.po' % (prj.slug, lang.code)
+
+        store = pofile()
+
+        site = Site.objects.get_current()
+        store.updateheader(
+            add = True,
+            language = lang.code,
+            x_generator = 'Weblate %s' % weblate.VERSION,
+            project_id_version = '%s dictionary for %s' % (lang.name, prj.name),
+            language_team = '%s <http://%s%s>' % (
+                lang.name,
+                site.domain,
+                reverse('weblate.trans.views.show_dictionary', kwargs = {'project': prj.slug, 'lang': lang.code}),
+            )
+        )
+
+        for word in words.iterator():
+            unit = pounit(word.source)
+            unit.target = word.target
+            store.addunit(unit)
+
+        store.savefile(response)
+
+        return response
 
 def show_dictionary(request, project, lang):
     prj = get_object_or_404(Project, slug = project)
@@ -1080,12 +1138,22 @@ def get_dictionary(request, unit_id):
     # join both lists
     words = set(words_std).union(words_stem)
 
+    # Grab all words in the dictionary
+    dictionary = Dictionary.objects.filter(
+        project = unit.translation.subproject.project,
+        language = unit.translation.language
+    )
+
+    # Built the query (can not use __in as we want case insensitive lookup)
+    query = Q()
+    for word in words:
+        query |= Q(source__iexact = word)
+
+    # Filter dictionary
+    dictionary = dictionary.filter(query)
+
     return render_to_response('js/dictionary.html', RequestContext(request, {
-        'dictionary': Dictionary.objects.filter(
-            project = unit.translation.subproject.project,
-            language = unit.translation.language,
-            source__in = words
-        ),
+        'dictionary': dictionary,
     }))
 
 @login_required